[TASK] Make the `CreateProcess` factory based on a interface
Basically, there are different ways to implement class
factories. Using a `final` class with a static method
without a interface on the factory class or the class
to be created by the factory is considerable bad and
prevents proper exchanging things.

A industry wide approach for providing good and exchangable
factories is following way:

* Build upon a factory class interface, ex. `MyFactoryInterface`
* Providing a abstract factory class implementing the interface
  and the default behaviour, ex. `AbstractFactory`
* Provide a slim class implementation based on the `AbstractFactory`,
  `MyFactory extends AbstractFactory`
* Add a DI configuration to retrieve the `MyFactory` based on the
  `MyFactoryInterface`.
* Use DI in code places to retrieve the factory based on the interface
  and operate with this.

To comply with this approach, this change ...

* adds a `CreateProcessFactoryInterface` with a non-static
  `get()` method.
* provide a abstract `AbstractCreateProcessFactory` class and
  move the code from the current `CreateProcessFactory` to the
  abstract implementation.
* Remove the method from the `CreateProcessFactory` and extend
  directly from the `AbstractCreateProcessFactory`
* Use the interface to retrieve the wanted factory implementation
  where the factory is required to retrieve the `CreateProcess`
  instance.

That allows consumer extension to create easier test mocks and
more important influence the instanciating in a variaty of ways.

Note: This is a more conveniant approach than adding a couple
      of limiting events to the factory class itself.

The adjusted DI approach for the factory class is marked as
internal and not considered as public api and can change at
any point in time.

Resolves: #9