Add API_FILEHOST to REST wizard for alternative file fetch
This commit adds an API_FILEHOST attribute to the RestWizardRequest class and updates all methods utilizing the API_HOST for fetching files to use this new attribute instead. This change provides the ability to specify an alternative location for fetching files, falling back to the API_URL if the API_FILEHOST is not set.