Breaking Change: Added CreateProcess connection to SourceInterface
The commit introduces a breaking change where a connection to the CreateProcess has been defined from SourceInterface. This addition requires sources to have a connection to the CreateProcess object. This is part of ongoing efforts to improve the process handling system. Please check the updated README for the code-snippet required to upgrade your source accordingly.