[TASK] create a user session
Some action one might take during a clone run might need a backend user session. A concrete Example would be renaming a folder using the FAL Framework during which a notice is generated and stored in the users session. While useful in the Backend, this would lead to an error if we don't have a dummy session in the CLI